feat: created the method to update task
The method updates task attributes and returns the fields that were
updated. It accepts one required argument and flexible arguments. This
means the user can pass in as many arguments as they want.

The taskID is required, and the attributes that the user would like to
update will pass in the method as flexible arguments. Since we don't
know what the user would give in, the method needs to validate inputs.